Week 6 Homework
Submitting all files for week 6 homework.  Do not be confused by the assignment_04_Adams_Kimberly.R that is both in the main Completed folder and in the Week 6 folder.  They are 2 different files.  The one in the Completed folder is homework from a different week that I mislabeled.  Haven't change it due to links in my submitted homework files.